minor logic change

+++ recipes/cudatoolkit/build.py
@@ -265,9 +265,11 @@ class WindowsExtractor(Extractor):
                             if not Path(os.path.join(store, filename)).is_file():
                                 shutil.copy(os.path.join(path, filename), store)
                         for filename in fnmatch.filter(files, "*.lib"):
-                            if path.endswith('x64'):
-                                if not Path(os.path.join(store, filename)).is_file():
-                                    shutil.copy(os.path.join(path, filename), store)
+                            if (
+                                path.endswith("x64")
+                                and not Path(os.path.join(store, filename)).is_file()
+                            ):
+                                shutil.copy(os.path.join(path, filename), store)
                         for filename in fnmatch.filter(files, "*.bc"):
                             if not Path(os.path.join(store, filename)).is_file():
                                 shutil.copy(os.path.join(path, filename), store)
